+++ src/p2p/p2p_server.h
@@ -7,6 +7,7 @@
 // Boost
 #include <boost/asio/io_context.hpp>
 #include <boost/asio/ip/tcp.hpp>
+#include <boost/asio/steady_timer.hpp>
 #include <boost/system/error_code.hpp>
 
 #include "metrics/stats_collector.h"
@@ -41,11 +42,13 @@ class P2PServer : public std::enable_shared_from_this<P2PServer>,
  private:
   void DoAccept();
   void OnAccept(boost::system::error_code ec);
+  void RestartAccept(boost::system::error_code ec);
 
  private:
   boost::asio::io_context& ioc_;
   boost::asio::ip::tcp::acceptor acceptor_;
   boost::asio::ip::tcp::socket socket_;
+  boost::asio::steady_timer accept_retry_timer_;
 
   RTCManager* rtc_manager_;
   P2PServerConfig config_;